FDF_SubmissionShare/FDF/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==2.0.0.post0
idna==3.4
numpy==1.23.4
pandas==1.5.1
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
